+++ apps/app/src/views/routes/Settings/components/Profile.tsx
@@ -14,7 +14,7 @@ import { graphql } from 'apps/app/src/gql';
 import { formatAddress } from 'apps/app/src/utils';
 import { useMemo } from 'react';
 import InfoGrid from '../../../components/InfoGrid';
-import { useClinicalApiClient } from '../apollo';
+import { useClinicalApiClient } from '../../../../clinicalApollo';
 
 const profileQuery = graphql(/* GraphQL */ `
   query MeProfileQuery {
